refactor: split frontend api client
